// src/types/CategoryIdinterface.tsx
export interface CharacterDetails {
  name: string;
  title?: string;
  height?: string;
  director?: string;
  model?: string;
  designation?: string;
  climate?: string;
  hair_color?: string;
  skin_color?: string;
  birth_year?: string;
  gender?: string;
  eye_color?: string;
  producer?: string;
  release_date?: string;
  consumables?: string;
  starship_class?: string;
  vehicle_class?: string;
  language?: string;
  gravity?: string;
  terrain?: string;
  created?: string;
  edited?: string;
  detail: string;
}